pyproject.toml
[tool.poetry]
name = "data-quality"
version = "0.1.0"
description = ""
authors = ["Mateus Kasuya <mateusvbkasuya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
mkdocs = "^1.6.0"
mkdocs-mermaid2-plugin = {extras = ["test"], version = "^1.1.1"}
mkdocs-material = "^9.5.24"
mkdocstrings = {extras = ["python"], version = "^0.25.1"}
taskipy = "^1.12.2"
isort = "^5.13.2"
black = "^24.4.2"
pytest = "^8.2.1"
pandas = "^2.2.2"
tqdm = "^4.66.4"
duckdb = "^0.10.3"
pydantic = "^2.7.1"
pandera = {extras = ["io"], version = "^0.19.3"}
sqlalchemy = "^2.0.30"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"